convert isbn type as Long to activate validations on property

+++ src/test/java/com/example/arhsbookstore/service/BookServiceTest.java
@@ -31,8 +31,8 @@ public class BookServiceTest {
         Mockito.when(bookRepository.findAll()).thenReturn(createBooks());
 
         List<Book> allBooks = bookService.getAllBooks();
-        assertEquals(1, allBooks.get(0).getIsbn());
-        assertEquals(2, allBooks.get(1).getIsbn());
+        assertEquals(1l, (long)allBooks.get(0).getIsbn());
+        assertEquals(2l, (long)allBooks.get(1).getIsbn());
     }
 
     @Test
@@ -45,7 +45,7 @@ public class BookServiceTest {
         Mockito.when(bookRepository.findById(1l)).thenReturn(Optional.of(mockBook));
 
         Book book = bookService.getBookByIsbn(1l);
-        assertEquals(1l, book.getIsbn());
+        assertEquals(1l, (long)book.getIsbn());
     }
 
     @Test(expected = NoSuchElementException.class)
@@ -65,7 +65,7 @@ public class BookServiceTest {
         Mockito.when(bookRepository.save(mockBook)).thenReturn(mockBook);
 
         Book book = bookService.saveOrUpdate(mockBook);
-        assertEquals(1l, book.getIsbn());
+        assertEquals(1l, (long)book.getIsbn());
     }
 
     @Test
@@ -75,21 +75,21 @@ public class BookServiceTest {
         Mockito.when(bookRepository.saveAll(mockBooks)).thenReturn(mockBooks);
 
         List<Book> allBooks = bookService.addBooks(mockBooks);
-        assertEquals(1, allBooks.get(0).getIsbn());
-        assertEquals(2, allBooks.get(1).getIsbn());
+        assertEquals(1l, (long)allBooks.get(0).getIsbn());
+        assertEquals(2l, (long)allBooks.get(1).getIsbn());
     }
 
 
 
     private List<Book> createBooks(){
         Book testBook1=new Book();
-        testBook1.setIsbn(1);
+        testBook1.setIsbn(1l);
         testBook1.setName("Lord of the Rings");
         testBook1.setAuthor("J.R.R. Tolkien");
         testBook1.setPublisher("Random House USA Inc");
 
         Book testBook2=new Book();
-        testBook2.setIsbn(2);
+        testBook2.setIsbn(2l);
         testBook2.setName("Mary Poppins");
         testBook2.setAuthor("P.L. Travers");
         testBook2.setPublisher("Thames and Hudson Ltd");
